moved pages to the pages folder
